% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analy_MCPMod.R
\name{analy_MCPMod}
\alias{analy_MCPMod}
\title{Output the proof-of concept (PoC) testing and the target dose.}
\usage{
analy_MCPMod(
  sim.obj = NULL,
  dose = NULL,
  resp = NULL,
  data = NULL,
  Mods.obj.cand,
  S = NULL,
  type = "normal",
  Delta,
  selModel = "aveAIC",
  ...
)
}
\arguments{
\item{sim.obj}{A data frame generated by \code{sim_dat} (recommended) that contains a dose variable and a response variable (when \code{type} = "normal") that serve as the input of \code{MCPMod}.}

\item{dose}{Either a vector specifying dose values or names of variables in the data frame specified in \code{data} (when \code{type} = "normal"), or a vector of the dose levels generated by \code{gen_design} (when \code{type} = "general") that serve as the input of \code{MCPMod}.}

\item{resp}{Either a vector specifying response values or names of variables in the data frame specified in \code{data} (when \code{type} = "normal"), or a vector of the estimates at the doses (when\code{type} = "general")that serve as the input of \code{MCPMod}.}

\item{data}{Data frame containing the variables referenced in dose and resp. If \code{data} is not specified, it is assumed that \code{dose} and \code{resp} are vectors specifying dose and response values.}

\item{Mods.obj.cand}{An object of class \code{Mods} that is used to specify the candidate dose-response models. See \code{Mods} for details.}

\item{S}{The covariance matrix of \code{resp} when \code{type} = "general".}

\item{type}{Determines whether inference is based on MCPMod applied on normally distributed response variable (when \code{type} = "normal"), or estimates at the doses and their covariance matrix and degrees of freedom that are specified in \code{resp}, \code{S}, and \code{df} through Generalized MCPMod (when \code{type} = "general").}

\item{Delta}{A numeric value indicating the clinically meaningful effect size.}

\item{selModel}{An optional character vector specifying the model selection criterion for dose estimation. Possible values are "AIC", "maxT", and "aveAIC" (recommended). For \code{type} = "general", the "gAIC" is used.}

\item{...}{Additional arguments to \code{MCPMod}, for example, placAdj, alpha, alternative, etc.}
}
\value{
A data frame that contains results for PoC (Proof-of-Concept) and MED (Minimum Effective Dose).
}
\description{
This function will conduct MCPMod analysis and return the result of the PoC testing of whether there is a significant dose-response relationship and the target dose, e.g., the minimum effective dose based on the clinically meaningful effect size.
}
\details{
Author: Qing Zhao
}
\examples{
\dontrun{
# when the response variable is normal 
gd.obj <- gen_design(dls = c(0, 6, 12, 18, 30), ngrp.equal = 60)
mods.true <- DoseFinding::Mods(sigEmax = c(6.5, 2.2), doses = gd.obj$dls, placEff = 0, maxEff = 1)
sds <- c(0.2, 0.6, 0.6, 0.6, 0.6)
dat <- sim_dat(dls = gd.obj$dls, ntot = gd.obj$ntot, ngrp = gd.obj$ngrp, Mods.obj.true = mods.true, sds = sds)
# when the response variable is binomial
gd.obj <- gen_design(dls = c(0, 4, 7, 10), ntot = 300, ratio =  c(2, 1, 2, 2))
mods.true <- DoseFinding::Mods(emax = 10, doses = gd.obj$dls, placEff = -1, maxEff = 1)
dat <- sim_dat(dls = gd.obj$dls, ntot = gd.obj$ntot, ngrp = gd.obj$ngrp, Mods.obj.true = mods.true, type = "binomial")
}
}
