% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_MCPMod.R
\name{eval_MCPMod}
\alias{eval_MCPMod}
\title{Generate evaluation matrix for simulation results.}
\usage{
eval_MCPMod(res.obj, true.med, lower, upper)
}
\arguments{
\item{res.obj}{A data frame generated by \code{analy_MCPMod} that contains the PoC and MED.}

\item{true.med}{A numeric value indicating the assumed true MED based on the assumed true dose response relationship and the clinically meaningful effect size.}

\item{lower}{A numeric value indicating the lower limit of the range that is reasonable for the estimated MED to fall relative to the assumed true MED.}

\item{upper}{A numeric value indicating the upper limit of the range that is reasonable for the estimated MED to fall relative to the assumed true MED.}
}
\value{
A list of key variables that contains the PoC, MED, POS, under (probability of underestimating the MED beyond the lower bound), and over (probability of overestimating the MED beyond the upper bound).
}
\description{
This function will create key results summary including PoC (Proof-of-Concept), MED (Minimum Effective Dose), POS (Probability of Successfully Identifying MED), and etc.
}
\details{
Author: Qing Zhao
}
\examples{
\dontrun{
pos.mat <- unlist(eval_MCPMod(res.obj, true.med, lower = 0.8, upper = 1.2))
}
}
