% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dat.R
\name{sim_dat}
\alias{sim_dat}
\title{Generate a dataset with dose and response variable.}
\usage{
sim_dat(dls, ntot, ngrp, Mods.obj.true, type = "normal", sds = NULL)
}
\arguments{
\item{dls}{A variable generated by \code{gen_design} (recommended) or a vector that contains the dose levels of interest. MCPMod requires at least three active doses for the Mod step.}

\item{ntot}{A variable generated by \code{gen_design} (recommended) or a numeric value indicating the total sample size.}

\item{ngrp}{A variable generated by \code{gen_design} (recommended) or a numeric value indicating the number of subjects in each dose group. The vector requires to be the same length as the dose level vector \code{dls}.}

\item{Mods.obj.true}{An object of class \code{Mods} {DoseFinding} that is used to specify the assumed true dose-response model. See \code{Mods} for details.}

\item{type}{A character vector to specify the type of the response variable, either "normal" (default) or "binomial".}

\item{sds}{An optional numeric vector indicating the standard deviations of the different dose groups when the type of the response variable is "normal". The vector requires to be the same length as the dose level vector \code{dls}.}
}
\value{
A data frame that contains the simulated data with subject level dose levels information and response information based on the specified assumed true dose response model.
}
\description{
This function will create a data frame that contains a dose variable based on the dose design of interest and a normal or binomial response variable based on the specified assumed true dose response model.
}
\details{
Author: Qing Zhao
}
\examples{
\dontrun{
#when the response variable is normal 
gd.obj <- gen_design(dls = c(0, 6, 12, 18, 30), ngrp.equal = 60)
mods.true <- DoseFinding::Mods(sigEmax = c(6.5, 2.2), doses = gd.obj$dls, placEff = 0, maxEff = 1)
sds <- c(0.2, 0.6, 0.6, 0.6, 0.6)
dat <- sim_dat(dls = gd.obj$dls, ntot = gd.obj$ntot, ngrp = gd.obj$ngrp, Mods.obj.true = mods.true, sds = sds)
# when the response variable is binomial
gd.obj <- gen_design(dls = c(0, 4, 7, 10), ntot = 300, ratio =  c(2, 1, 2, 2))
mods.true <- DoseFinding::Mods(emax = 10, doses = gd.obj$dls, placEff = -1, maxEff = 1)
dat <- sim_dat(dls = gd.obj$dls, ntot = gd.obj$ntot, ngrp = gd.obj$ngrp, Mods.obj.true = mods.true, type = "binomial")
}
}
